pyproject.toml
[tool.poetry]
name = "wp-bot"
version = "0.1.0"
description = ""
authors = ["fawolfmann <fawolfmann@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
WPP-Whatsapp = "^0.4.7"
requests = "^2.32.3"
faiss-cpu = "^1.8.0.post1"
sentence-transformers = "^3.0.1"
langchain = "^0.2.15"
llama-cpp-python = "^0.2.90"
chromadb = "^0.5.5"
llama-index = "^0.11.3"
llama-index-vector-stores-chroma = "^0.2.0"
llama-index-embeddings-huggingface = "^0.3.1"
llama-index-llms-llama-cpp = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
